!B(Odin Serializer) Distinguish the respective types, no longer Integer and Float to represent all types This will cause after the Weak is used for serialization, when deserializing, the generic interface is used to specify the exact type such as float, int and other non-double or long types. At this time, a conversion failure exception will occur.
`https://stackoverflow.com/questions/24447387/cast-object-containing-int-to-float-results-in-invalidcastexception`

+++ CabinIcarus.OdinSerializer/Core/Serializers/ByteSerializer.cs
@@ -36,7 +36,7 @@ namespace CabinIcarus.OdinSerializer
             string name;
             var entry = reader.PeekEntry(out name);
 
-            if (entry == EntryType.Integer)
+            if (entry == EntryType.Byte)
             {
                 byte value;
                 if (reader.ReadByte(out value) == false)
